did fractional knapsack problem. math looks good but I'll check later

+++ Assignment4/spice.hpp
@@ -8,7 +8,7 @@ using std::string;
 struct Spice {
 	string name;
 	float total_price;
-	int quantity;
+	float quantity;
 };
 
 #endif //SPICE_HPP
